==> 1_KEGG_download.ipynb <==
# Download global KEGG pathway database

import os
if not os.path.exists('KEGG'):
    os.makedirs('KEGG')

##  Download KEGG pathway id/names

import requests
from lxml import etree
import json

pathways_name_dir = {}

url = 'https://www.kegg.jp/kegg/pathway.html'
headers = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.54'}
html = requests.get(url,headers = headers)
e = etree.HTML(html.text)
idlist = e.xpath('//div[@class="list"]//dt/text()')
print('%d pathways ID found.'%len(idlist))
namelist = e.xpath('//div[@class="list"]//dd/a[1]/text()')
print('%d pathways name found.'%len(namelist))
if len(idlist) == len(namelist):
    for i in range(len(idlist)):
        pathways_name_dir['map'+idlist[i].strip()] = namelist[i].strip()
else:
    print('Error!')

with open('KEGG\\KEGG_pathways_name.json', 'w') as f:
    f.write(json.dumps(pathways_name_dir))
f.close()

## Download KO, EC, R entries in KEGG pathways

import pandas as pd
import requests
from lxml import etree
import json

with open('KEGG\\KEGG_pathways_name.json', 'r') as f:
    pathways_name_dir = json.load(f)
f.close()

pathways_list = [i for i in pathways_name_dir.keys()]

KEGG_pathways_entrys = pd.DataFrame(columns=['pw_ID','pw_name','coords','href','title'])

headers = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.54'}
m = 1
for pathway in pathways_list:
    pathway_name = pathways_name_dir[pathway]
    KEGG_pathway = requests.get('https://www.kegg.jp/pathway/'+pathway,headers = headers)
    e = etree.HTML(KEGG_pathway.text)
    hreflist = e.xpath('//map[@id="mapdata"]/area/@href')
    coordslist = e.xpath('//map[@id="mapdata"]/area/@coords')
    entrylist = e.xpath('//map[@id="mapdata"]/area/@title')
    print('found %d entries in %s'%(len(hreflist),pathway))
    if not len(hreflist) == len(entrylist):
        print(pathway,'Error!')
        continue
    for i in range(len(hreflist)):
        href = hreflist[i]
        coords = coordslist[i]
        title = entrylist[i]
        KEGG_pathways_entrys.loc[len(KEGG_pathways_entrys.index)]=[pathway,pathway_name,coords,href,title]

    print('map:'+pathway+'done!    '+str(m)+' / '+str(len(pathways_list)))
    m=m+1

KEGG_pathways_entrys.to_csv('KEGG\\KEGG_pathways_entrys.txt', header=True, index=False,sep='\t')

## Read R entries

import pandas as pd
import re

KEGG_pathways_reactions = pd.DataFrame(columns=['pw_ID','pw_name','coords','href','title','R_ID'])
KEGG_pathways_entrys = pd.read_csv('KEGG\\KEGG_pathways_entrys.txt', sep='\t',dtype={'pw_ID':str,'pw_name':str,'href':str,'title':str}).drop_duplicates()
for index, row in KEGG_pathways_entrys.iterrows():
    rs = re.findall(r'R\d{5}',row['title'])
    if rs:
        for r in rs:
            KEGG_pathways_reactions.loc[len(KEGG_pathways_reactions.index)]=[row['pw_ID'],row['pw_name'],row['coords'],row['href'],row['title'],r]
     
KEGG_pathways_reactions.drop_duplicates().reset_index().to_csv('KEGG\\KEGG_pathways_reactions.txt',sep='\t',index=False)

## Read KO entries

import pandas as pd
import re

KEGG_pathways_kos = pd.DataFrame(columns=['pw_ID','pw_name','coords','href','title','KO_ID'])
KEGG_pathways_entrys = pd.read_csv('KEGG\\KEGG_pathways_entrys.txt', sep='\t',dtype={'pw_ID':str,'pw_name':str,'href':str,'title':str}).drop_duplicates()
for index, row in KEGG_pathways_entrys.iterrows():
    kos = re.findall(r'K\d{5}',row['title'])
    if kos:
        for ko in kos:
            KEGG_pathways_kos.loc[len(KEGG_pathways_kos.index)]=[row['pw_ID'],row['pw_name'],row['coords'],row['href'],row['title'],ko]
     
KEGG_pathways_kos.drop_duplicates().reset_index().to_csv('KEGG\\KEGG_pathways_kos.txt',sep='\t',index=False)

## Download map picture

import pandas as pd
import requests
import json

with open('KEGG\\KEGG_pathways_name.json', 'r') as f:
    pathways_name_dir = json.load(f)
f.close()

pathways_list = [i for i in pathways_name_dir.keys()]

headers = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.54'}

m = 1
for pathway in pathways_list:
    downloadurl = 'https://www.kegg.jp/kegg/pathway/map/'+pathway+'@2x.png'
    headers={'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.54'}
    pict=requests.get(downloadurl,headers = headers).content
    with open('KEGG\\maps\\maps\\'+pathway+'.png','wb') as f:
        f.write(pict)
    print(pathway+' done!    '+str(m)+' / '+str(len(pathways_list)))
    m=m+1

## show reaction on map picture

import myModules.KEGG2BIGG as K2B
unikegglist = list(set(K2B.KEGG_pathways_reactions['R_ID'].tolist()))
a = K2B.pd.read_csv('KEGG\\Reaction_in_pathway_count.csv',index_col=0,squeeze=True)

from PIL import Image, ImageFont, ImageDraw

def drawtext(draw,text,coords_list):
    box = draw.textbbox((0,0),text, font=font)
    w = box[2] - box[0]
    h = box[3] - box[1]
    if type(coords_list) == list:
        coords_num = [2*int(i) for i in coords_list]
        x = (coords_num[0]+coords_num[2])/2
        y = (coords_num[1]+coords_num[3])/2
        draw.rectangle(coords_num, fill=(237, 245, 249), outline=(0,0,0),width=1)
        draw.text((x-w/2,y-h/2), text, font=font,fill=(0,0,0))
    else:
        print('coords_list should be a list')

map_list=[key for key in a.keys()]
#map_list = ['map00030']
font = ImageFont.truetype("arial.ttf", 18)
for mapp in map_list:
    map_reactions = K2B.KEGG_pathways_reactions[K2B.KEGG_pathways_reactions['pw_ID']==mapp]
    img = Image.open("KEGG\\maps\\maps\\%s.png"%mapp)
    draw = ImageDraw.Draw(img)
    for r in map_reactions['R_ID']:
        text = r
        findresult = map_reactions[map_reactions['R_ID']==r]['coords'].to_list()
        for i in findresult:
            coords_list = i.split(',')
            if len(coords_list) == 4:
                drawtext(draw,text,coords_list)
    img.save('KEGG\\maps\\mapasR\\%s_asR.png'%mapp)

# Download KEGG reaction/compound database

# download KEGG reactions
import requests
from lxml import etree
import re
import time
import pandas as pd

KEGG_REACTIONS_list = requests.get('https://rest.kegg.jp/list/reaction')
f = open("KEGG\\KEGG_REACTIONS_list.txt", 'w')
f.write(KEGG_REACTIONS_list.text)

kegg_rxn_list = pd.read_table('KEGG\\KEGG_REACTIONS_list.txt',header=None)
kegg_rlist = [st.replace('rn:','') for st in kegg_rxn_list[0]]

headers = {'User-Agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36 Edg/108.0.1462.54'}
f = open("KEGG\\KEGG_REACTIONS_DB.txt", 'a')
f.write('ID\tNAMES\tKEGG\tEQUATION\tECs\n')

for id in kegg_rlist:
    url ='https://www.kegg.jp/entry/'+id
    try:
        html = requests.get(url,headers = headers)
    except requests.exceptions.Timeout:
        print("time out, try again")
        time.sleep(20)
        html = requests.get(url,headers = headers)
    if not html.ok:
        print('wrong reaction ID')
        time.sleep(20)
        html = requests.get(url,headers = headers)
        if not html.ok:
            continue

    e = etree.HTML(html.text)

    name_content = e.xpath('//th/span[text()="Name"]/../../td/div/div/text()')
    name = ''
    for st in name_content[:-1]:
        name = name +st.replace('\n','').replace(';','|')
    
    equation_content = e.xpath('//th/span[text()="Equation"]/../../td/div//text()')
    equation = ''
    for st in equation_content[:-1]:
        if st.startswith('C'):
            st = st+'[0]'
        else:
            matches = re.compile(r"\d+").findall(st)
            if matches:
                match = matches[0]
                st = st.replace(match, f"({match})")    
        equation = equation + st
    
    EC_content = e.xpath('//th/span[text()="Enzyme"]/../../td/div//text()')
    EC1 = [i.replace(u'\xa0','\n') for i in EC_content]
    EC2 = [i for i in EC1 if not str(i).startswith('\n')]
    EC = ''
    for st in EC2:
        EC = EC + st + '|'
    EC = EC[:-1]
    f.write(id+'\t'+name+'\t'+id+'\t'+equation+'\t'+EC+'\n')
f.close()


# KEGG reaction to ModelSEED reaction
import modelseedpy

modelseed_path = 'C:\\Users\\vickenlee\\ModelSEEDDatabase'
modelseed = modelseedpy.biochem.from_local(modelseed_path)

df_keggID_modelseedID = pd.DataFrame(columns=["Kegg_ID","ModelSeed_ID"])
for (key,value) in modelseed.reaction_aliases.items():
    ModelSeed_ID = key
    try:    
        Kegg_ID = list(value.get('KEGG'))[0]
    except:
            Kegg_ID =''
    df_keggID_modelseedID.loc[len(df_keggID_modelseedID.index)]=[Kegg_ID,ModelSeed_ID]
df_keggID_modelseedID.to_csv('KEGG\\KEGG_reaction_keggID_modelseedID.txt',  header=True, index=False,sep='\t')

print('done!')

# download KEGG compounds and map to ModelSEED compounds

KEGG_COMPOUNDS_list = requests.get('https://rest.kegg.jp/list/compound')
f = open("KEGG\\KEGG_COMPOUNDS_list.txt", 'w')
f.write(KEGG_COMPOUNDS_list.text)
kegg_cpd_list = pd.read_table('KEGG\\KEGG_COMPOUNDS_list.txt',header=None)
KEGG_COMPOUNDS_name = pd.DataFrame(columns=['ID','Name'])
KEGG_COMPOUNDS_name['ID']=[st.replace('cpd:','') for st in kegg_cpd_list[0]]
KEGG_COMPOUNDS_name['Name']=[st.replace('; ','|') for st in kegg_cpd_list[1]]


import modelseedpy
modelseed_path = 'C:\\Users\\vickenlee\\ModelSEEDDatabase'
modelseed = modelseedpy.biochem.from_local(modelseed_path)
cpd_keggID_modelseedID = pd.DataFrame(columns=["Kegg_ID","ModelSeed_ID","formula","charge"])
for (key,value) in modelseed.compound_aliases.items():
    ModelSeed_ID = key
    compound = modelseed.get_seed_compound(ModelSeed_ID)
    formula = compound.formula
    charge = compound.data['charge']
    try:    
        find_Kegg = list(value.get('KEGG'))
    except:
            find_Kegg =''
    if find_Kegg:
        for Kegg_ID in find_Kegg:
            cpd_keggID_modelseedID.loc[len(cpd_keggID_modelseedID.index)]=[Kegg_ID,ModelSeed_ID,formula,charge]
cpd_keggID_modelseedID.to_csv('KEGG\\KEGG_compound_keggID_modelseedID.txt', header=True,sep='\t',index=False)
cpd_keggID_modelseedID=cpd_keggID_modelseedID.rename(columns={'Kegg_ID':'ID'})


KEGG_COMPOUNDS_DB = pd.merge(KEGG_COMPOUNDS_name,cpd_keggID_modelseedID,how='left',on='ID').drop_duplicates(subset=['ID'], keep='first')


from lxml import etree
for index, row in KEGG_COMPOUNDS_DB.iterrows():
    if not row['formula']==row['formula']:
        url ='https://www.kegg.jp/entry/'+row['ID']
        try:
            html = requests.get(url,headers = headers)
        except requests.exceptions.Timeout:
            time.sleep(20)
            html = requests.get(url,headers = headers)
        e = etree.HTML(html.text)
        formula_content = e.xpath('//th/span[text()="Formula"]/../../td/div/text()')
        if formula_content:
            formula = formula_content[0]
            KEGG_COMPOUNDS_DB['formula'][index]=formula
            print(row['ID']+': add formula')
        else:
            print(row['ID']+': does not have formula')

KEGG_COMPOUNDS_DB.to_csv('Database\\KEGG_COMPOUNDS_DB.txt',sep='\t',index=None)

# Download species KEGG

org_name = 'mko'
org_full_name = 'Methylomonas_koyamae'

org_name = 'mbry'
org_full_name = 'Methylocystis_bryophila'

org_name = 'mca'
org_full_name = 'Methylococcus_capsulatus'

import myModules.KEGGdownload as kd

pwlist = kd.getpwlist(org_name)

kd.downloadxml(org_name,pwlist)

df_reaction = kd.getentriesfromxml(org_name)

# Get KEGG reaction from genome

org_name = 'mko'
org_full_name = 'Methylomonas_koyamae'

org_name = 'mbry'
org_full_name = 'Methylocystis_bryophila'

org_name = 'mca'
org_full_name = 'Methylococcus_capsulatus'

import os
faa_paths=[]
for root, dirs, files in os.walk('genome\\'+org_full_name+'_genome'):
    if 'protein.faa' in files:
        genome_path = root+'\\protein.faa'
print(genome_path)


1. upload protein.faa to https://www.kegg.jp/blastkoala/

2. download the blast result to genome\\XXX_genome\\user_ko.txt

import myModules.KEGG2BIGG as K2B
import pandas as pd

df_org_ko = pd.read_csv('genome\\'+org_full_name+'_genome\\user_ko.txt',sep='\t',header=None)
df_org_ko.columns = ['Genes','KO_ID']
ko_list = df_org_ko['KO_ID'].tolist()
R_list =[]
for ko in ko_list:
    try:
        R = K2B.getRfromKO(ko)
        if type(R) == list:
            R_list+=R
        elif type(R) == str:
            R_list.append(R)
    except:
        continue
R_list = list(set(R_list))
ko_list = list(set([i for i in ko_list if i==i and not i == 'ko']))
with open('genome\\'+org_full_name+'_genome\\R_list.txt','w') as f:
    for R in R_list:
        f.write(str(R)+'\n')
f.close()
with open('genome\\'+org_full_name+'_genome\\ko_list.txt','w') as f:
    for ko in ko_list:
        f.write(str(ko)+'\n')

### show reaction on map picture

import myModules.KEGG2BIGG as K2B
with open('genome\\'+org_full_name+'_genome\\ko_list.txt','r') as f:
    ko_list = f.read().splitlines()
f.close()
a = K2B.pathway_count(ko_list)

from PIL import Image, ImageFont, ImageDraw

def drawtext(draw,coords_list):
    if type(coords_list) == list:
        coords_num = [2*int(i) for i in coords_list]
        draw.rectangle(coords_num, fill=None, outline=(252, 85, 49),width=5)
    else:
        print('coords_list should be a list')

map_list=[key for key in a.keys()]
#map_list = ['map00680']
font = ImageFont.truetype("arial.ttf", 18)
for mapp in map_list:
    map_kos = K2B.KEGG_pathways_kos[K2B.KEGG_pathways_kos['pw_ID']==mapp]
    try:
        img = Image.open("KEGG\\maps\\mapedinBIGGandR\\%s_biggandR.png"%mapp)
    except:
        continue
    draw = ImageDraw.Draw(img)
    for ko in ko_list:
        text = ko
        findresult = map_kos[map_kos['KO_ID']==ko]['coords'].to_list()
        for i in findresult:
            coords_list = i.split(',')
            if len(coords_list) == 4:
                drawtext(draw,coords_list)
    img.save('KEGG\\maps\\%s_maps\\%s_%s.png'%(org_name,mapp,org_name))
