==> 3_Building_model.ipynb <==
org_name = 'mko'
org_full_name = 'Methylomonas_koyamae'

org_name = 'mbry'
org_full_name = 'Methylocystis_bryophila'

org_name = 'mca'
org_full_name = 'Methylococcus_capsulatus'

## import KEGG reaction list and mapping to BIGG reaction

import myModules.KEGG2BIGG as K2B

unikegglist = list(set(K2B.KEGG_pathways_reactions['R_ID'].tolist()))
unikeggmappedinBIGG = [r for r in unikegglist if K2B.KEGG_BIGG_dir_selected.get(r)]
unikeggnotmappedinBIGG = [r for r in unikegglist if not K2B.KEGG_BIGG_dir_selected.get(r)]

import pandas as pd
df_kegg_reaction  = pd.read_excel(org_name+'\\'+org_name+'_clean_kegg_reaction.xlsx', index_col=0) 
keggrxnlist=df_kegg_reaction['Kegg_ID'].tolist()

R_list = []
with open('genome\\'+org_full_name+'_genome\\R_list.txt','r') as f:
    for line in f.readlines():
        R_list.append(line.strip())
f.close()

print('kegg_rxn_list: '+str(len(keggrxnlist)))
print('R_list: '+str(len(R_list)))

mcakeggrxnlist=pd.read_excel('mca\\mca_clean_kegg_reaction.xlsx', index_col=0) ['Kegg_ID'].tolist()

with open(org_name+'\\'+org_name+'_mca_inBIGG_KEGG_color.txt',"w") as f:
    kegglist = list(set(keggrxnlist+mcakeggrxnlist+unikeggmappedinBIGG))
    for r in kegglist:
        string = r
        if r in keggrxnlist:
            string+=' yellow'
        else:
            string+=' white'
        if r in mcakeggrxnlist:
            string+=' pink'
        else:
            string+=' white'
        if r in unikeggmappedinBIGG:
            string+=' green'
        else:
            string+=' white'
        f.write(string+'\n')
f.close()

# show color in website https://www.kegg.jp/kegg-bin/show_pathway?map00680 

mappedBIGG = {}
can_not_map = []
for r in keggrxnlist:
    if K2B.KEGG_BIGG_dir_selected.get(r):
        mappedBIGG[r] = K2B.KEGG_BIGG_dir_selected.get(r)
    else:
        can_not_map.append(r)
print('mappedBIGG: '+str(len(mappedBIGG)))
print('can_not_map: '+str(len(can_not_map)))

mappedKEGGlist = list(set(mappedBIGG.keys()))
mappedBIGGlist = list(set(mappedBIGG.values()))
print('mappedBIGGlist: '+str(len(mappedBIGGlist)))
cannotmaplist = list(set(can_not_map))
print('cannotmaplist: '+str(len(cannotmaplist)))

import pandas as pd
uniR_in_pw_count = pd.read_csv('KEGG\\Reaction_in_pathway_count.csv',index_col=0,squeeze=True)

R_in_pw_count = K2B.pathway_count(keggrxnlist)

unbigged_r_in_pw_count = K2B.pathway_count(cannotmaplist,R_in_pw_count)

bigged_r_in_pw_count = K2B.pathway_count(mappedKEGGlist,R_in_pw_count)

## get mapped Excel

import myModules.buildmodel as BM
import cobra

mapped_model = cobra.Model('mapped_in_BIGG')

BM.add_rxn_from_unibigg(mapped_model,mappedBIGGlist)

mapped_model

def find_key (dict, value):
    keys = [k for k, v in dict.items() if v == value]
    if len(keys)==1:
        return keys[0]
    else:
        print(value + ' is not unique')
        return keys[0]

import pandas as pd
df_reaction = pd.DataFrame(columns=['ID','Name','Reaction','KEGG_ID','Pathway'])
for rxn in mapped_model.reactions:
    keggid = find_key(mappedBIGG,rxn.id)
    df_reaction.loc[len(df_reaction)] = [rxn.id,rxn.name,rxn.reaction,keggid,K2B.getPathways(keggid)]
df_reaction.to_excel(org_name+'\\01_'+org_name+'_export_mapped_reaction.xlsx',index=False)

df_cannotmappedreaction = pd.DataFrame(columns=['KEGG_ID','Pathway'])
for rxn in cannotmaplist:
    df_cannotmappedreaction.loc[len(df_cannotmappedreaction)] = [rxn,K2B.getPathways(rxn)]
df_cannotmappedreaction.to_excel(org_name+'\\02_'+org_name+'_export_cannotmapped_reaction.xlsx',index=False)

## Adjust by excel

copy 01_export_mapped_reaction.xlsx as 03_import_reaction.xlsx

import myModules.buildmodel as BM

BM.print_rxn(BM.biggunimodel,'UGLDDS1','UAAGLS2')

BM.print_metabolite(BM.biggunimodel,'udcpp_c')

BM.print_metabolite(BM.biggunimodel,'frdp_c')

BM.biggunimodel.metabolites.get_by_id('octdp_c')

rr = BM.biggunimodel.metabolites.get_by_id('udcpdp_c')
for r in rr.reactions:
    BM.print_rxn(BM.biggunimodel,r.id)

r1 = BM.biggunimodel.metabolites.get_by_id('thmmp_c').reactions
r2 = BM.biggunimodel.metabolites.get_by_id('2mahmp_c').reactions
for r in r1:
    if r in r2:
        BM.print_rxn(BM.biggunimodel,r.id)

K2B.get_pathway_from_bigg('FBA')

## Add reactions from Excel

import myModules.buildmodel as BM

import cobra
new_model = cobra.Model(org_full_name)
new_model

import pandas as pd
df_import_reaction = pd.read_excel(org_name+'\\03_'+org_name+'_import_reaction.xlsx')
for i in range(len(df_import_reaction)):
    rxn_id = df_import_reaction.iloc[i]['ID']
    rxn_name = df_import_reaction.iloc[i]['Name']
    rxn_string = df_import_reaction.iloc[i]['Reaction']
    BM.add_rxn_by_string(new_model,rxn_id,rxn_name,rxn_string)


check these not in BIGG metabolites. list 'ID, Name, KEGG' in the import_reactions.xlsx Sheet2

import json
seedCOMDB = json.load(open('.\\SEED\\compounds.json','r'))
def add_to_dir(my_dir, key, value):
    if key in my_dir:
        if isinstance(my_dir[key], list) and value not in my_dir[key]:
            my_dir[key].append(value)
        else:
            my_dir[key] = [my_dir[key], value]
    else:
        my_dir[key] = value

def find_compound_by_seedid(id,seedCOMDB=seedCOMDB):
    for c in seedCOMDB:
        if c['id'] == id:
            return c
    return None

def find_compound_by_keggid(id,seedCOMDB=seedCOMDB):
    for c in seedCOMDB:
        if c.get('aliases'):
            for l in c['aliases']:
                if l.startswith('KEGG:'):
                    for s in l.split(':')[1].split(';'):
                        if s.strip() == id:
                            return c
    return None

def get_anno_from_seed(c):
    anno_dict = {}
    add_to_dir(anno_dict,'seed.compound',c['id'])
    if c.get('inchikey'):
        add_to_dir(anno_dict,'inchikey',c['inchikey'])
    for l in c['aliases']:
        if l.startswith('KEGG:'):
            for s in l.split(':')[1].split(';'):
                add_to_dir(anno_dict,'kegg.compound',s.strip())
        elif l.startswith('MetaCyc:'):
            for s in l.split(':')[1].split(';'):
                add_to_dir(anno_dict,'biocyc','META:'+s.strip())
    return anno_dict

df_import_metabolite = pd.read_excel(org_name+'\\03_'+org_name+'_import_reaction.xlsx',sheet_name='Sheet2')
for i in range(len(df_import_metabolite)):
    met_id = df_import_metabolite.iloc[i]['ID']
    met_name = df_import_metabolite.iloc[i]['Name']
    met_kegg_id = df_import_metabolite.iloc[i]['KEGG']
    new_model.metabolites.get_by_id(met_id).name = met_name
    c = find_compound_by_keggid(met_kegg_id)
    if c:
        new_model.metabolites.get_by_id(met_id).formula = c['formula']
        new_model.metabolites.get_by_id(met_id).charge = c['charge']
        new_model.metabolites.get_by_id(met_id).annotation = get_anno_from_seed(c)
    else:
        new_model.metabolites.get_by_id(met_id).formula = met_kegg_id
        new_model.metabolites.get_by_id(met_id).charge = 0

### Add annotation

for rxn in new_model.reactions:
    BM.add_annotation_from_BIGGurl(rxn)

check these not in BIGG reactions. list 'ID, KEGG_R' in the import_reactions.xlsx Sheet3

import json
seedrxnDB = json.load(open('.\\SEED\\reactions.json','r'))
def add_to_dir(my_dir, key, value):
    if key in my_dir:
        if isinstance(my_dir[key], list) and value not in my_dir[key]:
            my_dir[key].append(value)
        else:
            my_dir[key] = [my_dir[key], value]
    else:
        my_dir[key] = value

def find_reaction_by_seedid(id,seedrxnDB=seedrxnDB):
    for r in seedrxnDB:
        if r['id'] == id:
            return r
    return None

def find_reaction_by_keggid(id,seedrxnDB=seedrxnDB):
    for r in seedrxnDB:
        if r.get('aliases'):
            for l in r['aliases']:
                if l.startswith('KEGG:'):
                    for s in l.split(':')[1].split(';'):
                        if s.strip() == id:
                            return r
    return None

def get_anno_from_seedrxn(r):
    anno_dict = {}
    add_to_dir(anno_dict,'seed.reaction',r['id'])
    if r.get('ec_numbers'):
        for ec in r['ec_numbers']:
            add_to_dir(anno_dict,'ec-code',ec)
    for l in r['aliases']:
        if l.startswith('KEGG:'):
            for s in l.split(':')[1].split(';'):
                add_to_dir(anno_dict,'kegg.reaction',s.strip())
        elif l.startswith('MetaCyc:'):
            for s in l.split(':')[1].split(';'):
                add_to_dir(anno_dict,'biocyc','META:'+s.strip())
    return anno_dict

df_import_reaction_anno = pd.read_excel(org_name+'\\03_'+org_name+'_import_reaction.xlsx',sheet_name='Sheet3')
for i in range(len(df_import_reaction_anno)):
    rxn_id = df_import_reaction_anno.iloc[i]['ID']
    rxn_kegg_id = df_import_reaction_anno.iloc[i]['KEGG_R']
    r = find_reaction_by_keggid(rxn_kegg_id)
    if r:
        new_model.reactions.get_by_id(rxn_id).annotation = get_anno_from_seedrxn(r)
    else:
        new_model.reactions.get_by_id(rxn_id).annotation = {'kegg.reaction':rxn_kegg_id}

for rxn in new_model.reactions:
    if rxn.annotation.get('kegg.reaction') == None:
        try:
            keggid = df_import_reaction[df_import_reaction["ID"]==rxn.id]["KEGG_R"].to_list()[0].split(',')
            rxn.annotation['kegg.reaction'] = keggid
        except:
            pass

for rxn in new_model.reactions:
    if rxn.annotation.get('kegg.reaction') == None:
        print(rxn.id)

loca = ['c','p','e']
for l in loca:
    new_model.metabolites.get_by_id('h2o_'+l).name = 'H2O'
    new_model.metabolites.get_by_id('o2_'+l).name = 'O2'
    new_model.metabolites.get_by_id('co2_'+l).name = 'CO2'
    new_model.metabolites.get_by_id('h_'+l).charge=1

BM.add_rxn_by_string(new_model,'H2tce','H2 transport','h2_e <=> h2_c')
BM.add_rxn_by_string(new_model,'cobalt2tce','Cobalt2 transport','cobalt2_e <=> cobalt2_c')
BM.add_rxn_by_string(new_model,'Cutce','Cu transport','cu2_e <=> cu2_c')
BM.add_rxn_by_string(new_model,'Fetce','Fe transport','fe2_e <=> fe2_c')

new_model

### Save excel reactions

import pandas as pd
import json
df_ex_reaction = pd.DataFrame(columns=['ID','Name','Reaction','Annotation'])
for rxn in new_model.reactions:
    ID = rxn.id
    Name = rxn.name
    Reaction = rxn.reaction
    Annotation = json.dumps(rxn.annotation)
    df_ex_reaction.loc[len(df_ex_reaction)] = [ID,Name,Reaction,Annotation]
df_ex_reaction.to_excel(org_name+'\\04_'+org_name+'_export_clean_reaction.xlsx', index=False,sheet_name='Reactions')

df_ex_metabolite = pd.DataFrame(columns=['ID','Name','Formula','Charge','Annotation'])
for met in new_model.metabolites:
    ID = met.id
    Name = met.name
    Formula = met.formula
    Charge = met.charge
    Annotation = json.dumps(met.annotation)
    df_ex_metabolite.loc[len(df_ex_metabolite)] = [ID,Name,Formula,Charge,Annotation]
df_ex_metabolite.to_excel(org_name+'\\05_'+org_name+'_export_clean_metabolite.xlsx', index=False,sheet_name='Metabolites')

cobra.io.write_sbml_model(new_model, org_name+'\\06_'+org_name+'_excel_model.xml')
cobra.io.save_json_model(new_model, org_name+'\\06_'+org_name+'_excel_model.json')

## Add biomass reaction

### import other model and get the biomass reactions

GB1_model = cobra.io.read_sbml_model('other_model\\5GB1_base_fermentation.xml')

biorxn = GB1_model.reactions.get_by_id('BIOMASS_M5GB1')
for meta, sto in biorxn.metabolites.items():
    try:
        a = new_model.metabolites.get_by_id(meta.id)
        print(meta.id,'\t', sto)
    except:
        print(meta.id,'\t', sto,'\t',meta.name)

# import new mca model and get biomass reaction, then get list of metabolites, then compare to bigguni model
def find_kegg_com_in_uniBIGG(kegg_id):
    for m in BM.biggunimodel.metabolites:
        if m.annotation.get('kegg.compound'):
            if kegg_id in m.annotation.get('kegg.compound'):
                return m.id
mca_model = cobra.io.read_sbml_model('other_model\\newmethylococcus_new.xml')
mcabiorxn = mca_model.reactions.get_by_id('Biomass_reaction')
mcabiolist=[]
biodict = {}
for meta, sto in mcabiorxn.metabolites.items():
    meta_id = meta.id.split('_')[0]
    mcabiolist.append(meta_id)
    if find_kegg_com_in_uniBIGG(meta_id):
        biodict[find_kegg_com_in_uniBIGG(meta_id)] = sto
    else:
        print('not found in bigguni model:\t'+meta_id+'\t'+str(sto)+'\t'+meta.name)


for key, item in biodict.items():
    try:
        a=new_model.metabolites.get_by_id(key)
        print(key,'\t',item)
    except:
        print(key,'\t',item,'\t','not found in new model')

compare and make a bio rxn (id and sto) in import_reaction.xlsx sheet4

org_name = 'mko'
org_full_name = 'Methylomonas_koyamae'

import cobra
import pandas as pd
new_model = cobra.io.read_sbml_model(org_name+'\\06_'+org_name+'_excel_model.xml')

biomass_meta = cobra.Metabolite('mko_biomass_c')
biomass_meta.name = 'mko_biomass'
biomass_meta.compartment = 'c'
biomass_meta.formula = 'C41H75O22N9P'
biomass_meta.charge = -121

new_model.add_metabolites(biomass_meta)

new_model.metabolites.get_by_id('mko_biomass_c')

df_import_bio_reaction = pd.read_excel(org_name+'\\03_'+org_name+'_import_reaction.xlsx',sheet_name='Sheet4')
biomass_meta = {}
for i in range(len(df_import_bio_reaction)):
    met_id = df_import_bio_reaction.iloc[i]['meta_ID']
    met_sto = df_import_bio_reaction.iloc[i]['STO']
    biomass_meta[met_id] = met_sto
print(len(biomass_meta))

biomass_rxn = cobra.Reaction('Biomass_reaction')
biomass_rxn.name = org_full_name +' biomass reaction'
biomass_rxn.lower_bound = 0
biomass_rxn.upper_bound = 1000
for meta, sto in biomass_meta.items():
    try:
        biomass_rxn.add_metabolites({new_model.metabolites.get_by_id(meta):sto})
    except:
        print(meta+' not found in new model')

new_model.add_reactions([biomass_rxn])

len(new_model.reactions.get_by_id('Biomass_reaction').metabolites)

new_model.add_boundary(new_model.metabolites.get_by_id('mko_biomass_c'), type="demand")

new_model.objective = 'Biomass_reaction'

new_model.compartments= {'c':'cytosol','p':'periplasm','e':'extracellular space','m':'membrane'}

## Add exchange rxn

### Get other model exchanges

for rxn in mca_model.exchanges:
    print(rxn.id,'\t',rxn.name,'\t',rxn.reaction)

for rxnn in GB1_model.exchanges:
    print(rxnn.id,'\t',rxnn.name,'\t',rxnn.reaction)

### Add EX reactions

met_e = []
for met in new_model.metabolites:
    if met.id.endswith('_e'):
        new_model.add_boundary(met, type="exchange")
        #print(met.id)

new_model.reactions.EX_meoh_e.lower_bound = 0
new_model.reactions.EX_pyr_e.lower_bound = 0
new_model.reactions.EX_ac_e.lower_bound = 0
new_model.reactions.EX_for_e.lower_bound = 0
new_model.reactions.EX_lac__D_e.lower_bound = 0
new_model.reactions.EX_succ_e.lower_bound = 0
new_model.reactions.EX_glyc_e.lower_bound = 0
new_model.reactions.EX_no2_e.lower_bound = 0
new_model.reactions.EX_n2o_e.lower_bound = 0
new_model.reactions.EX_no_e.lower_bound = 0
new_model.reactions.EX_nh4_e.lower_bound = 0

new_model.reactions.EX_ch4_e.lower_bound = -30
new_model.reactions.EX_ch4_e.upper_bound = 0
new_model.reactions.EX_o2_e.lower_bound = -45
new_model.reactions.EX_o2_e.upper_bound = -35

for rr in new_model.exchanges:
    print(rr.id,'\t',rr.name,'\t',rr.reaction)

cobra.io.write_sbml_model(new_model, org_name+'\\07_'+org_name+'_biomass_model.xml')
cobra.io.save_json_model(new_model, org_name+'\\07_'+org_name+'_biomass_model.json')

## Add sink rxn

meta_list = ["amp_c","CPC161161_c","cys__L_c","datp_c","sql_c","dgtp_c","CPC161140_c","udcpp_c","thr__L_c","CPC160161_c","ile__L_c","his__L_c","leu__L_c","dctp_c","nad_c","met__L_c"]
for meta in meta_list:
    new_model.add_boundary(new_model.metabolites.get_by_id(meta), type="sink")
    new_model.reactions.get_by_id('SK_'+meta).upper_bound = 0

for check_mete in new_model.metabolites:
    try:
        new_model.add_boundary(check_mete, type="sink")
    except:
        continue
    new_model.reactions.get_by_id('SK_'+check_mete.id).upper_bound = 0
    b= new_model.optimize()
    if b.objective_value > 0.001:
        print(check_mete.id+' is useful')
    new_model.remove_reactions([new_model.reactions.get_by_id('SK_'+check_mete.id)])


cobra.io.write_sbml_model(new_model, org_name+'\\08_'+org_name+'_complete_model_V1.xml')
cobra.io.save_json_model(new_model, org_name+'\\08_'+org_name+'_complete_model_V1.json')

## Export model

import pandas as pd
df_reaction = pd.DataFrame(columns=['ID','Name','Reaction','Pathway'])
for rxn in new_model.reactions:
    df_reaction.loc[len(df_reaction)] = [rxn.id,rxn.name,rxn.reaction,K2B.get_pathway_from_bigg(rxn.id)]
df_reaction.to_excel(org_name+'\\'+org_name+'_export_reaction_full.xlsx',index=False)


new_model.summary()
