==> 4_Gipfilling.ipynb <==
org_name = 'mko'
org_full_name = 'Methylomonas_koyamae'

org_name = 'mbry'
org_full_name = 'Methylocystis_bryophila'

org_name = 'mca'
org_full_name = 'Methylococcus_capsulatus'

import cobra
old_model = cobra.io.read_sbml_model(org_name+'\\08_'+org_name+'_complete_model_V1.xml')

a = old_model.optimize()
a.fluxes.to_csv(org_name+'\\08_'+org_name+'_origin_fluxes.csv')
old_model.summary()

old_model.objective = 'EX_ac_e'

old_model.optimize().fluxes.to_csv(org_name+'\\09_'+org_name+'_EX_ac_fluxes.csv')

old_model.remove_reactions([old_model.reactions.get_by_id('SK_amp_c')])

old_model.reactions.get_by_id('EX_o2_e').bounds = (-1000,0)

old_model.reactions.get_by_id('EX_co2_e').bounds = (0,20)

old_model.summary()

optimize_dict = {}
for rxn in old_model.reactions:
    old_model.objective = rxn
    a=old_model.optimize()
    optimize_dict[rxn.id] = a.objective_value

print(len(optimize_dict))
zero_rxn = []
for key, value in optimize_dict.items():
    if value == 0:
        zero_rxn.append(key)
print(len(zero_rxn))

for key, value in optimize_dict.items():
    print(key, value)

import myModules.buildmodel as BM

carveme_model = cobra.io.read_sbml_model(org_name+'\\automodel\\'+org_name+'_carveme_model.xml')

for mete in old_model.reactions.get_by_id("Biomass_reaction").metabolites:
    

carveme_model.summary()

imc_model = cobra.io.read_sbml_model('other_model\\iMcBath.xml')
imc_model

for me,sto in imc_model.reactions.BIOMASS_Mcapsulatus.metabolites.items():
    print(me.id,'\t',sto,'\t',me.name)

solution = cobra.flux_analysis.gapfill(old_model, imc_model, demand_reactions=False)

for reaction in solution[0]:
    print(reaction.id)

bigg_model = cobra.io.load_json_model('.\\BIGG\\universal_model_modified.json')

l2 = len(bigg_model.reactions)-10

def check_not_c(rxn):
    for met in rxn.metabolites:
        if not met.id.endswith('_c'):
            return True
    return False

end2 =3
end1 =2
print("time2: "+str((end2-end1))+"s")

with open('removel_bigg_model_reactions.txt', 'w') as f:
    for rxn in bigg_model.reactions:
        if check_not_c(rxn):
            f.write(rxn.id+'\n')
            bigg_model.remove_reactions([rxn])
f.close()

bigg_model
